refactor(native): extract sidecar lifecycle
